+++ cliquet/resource.py
@@ -459,8 +459,9 @@ class BaseResource(object):
             in the iterim.
         """
         self._raise_400_if_invalid_id(self.record_id)
-        self._add_timestamp_header(self.request.response)
         record = self._get_record_or_404(self.record_id)
+        timestamp = record[self.collection.modified_field]
+        self._add_timestamp_header(self.request.response, timestamp=timestamp)
         self._raise_304_if_not_modified(record)
         self._raise_412_if_modified(record)
 
@@ -519,6 +520,9 @@ class BaseResource(object):
                                                    unique_fields=unique_fields)
         except storage_exceptions.UnicityError as e:
             self._raise_conflict(e)
+
+        timestamp = record[self.collection.modified_field]
+        self._add_timestamp_header(self.request.response, timestamp=timestamp)
 
         body = {
             'data': record,
@@ -594,6 +598,10 @@ class BaseResource(object):
                     if changes.get(k) != new_record.get(k)}
         else:
             data = new_record
+
+        timestamp = new_record.get(self.collection.modified_field,
+                                   old_record[self.collection.modified_field])
+        self._add_timestamp_header(self.request.response, timestamp=timestamp)
 
         body = {
             'data': data,